// api/src/pocketbase-types.ts
export type IsoDateString = string
export type IsoAutoDateString = string & { readonly autodate: unique symbol }
export type RecordIdString = string
export type FileNameString = string & { readonly filename: unique symbol }
export type HTMLString = string
// ...
export const EpisodesStatusOptions = {
	"draft": "draft",
	"scheduled": "scheduled",
	"live": "live",
	"completed": "completed",
	"cancelled": "cancelled",
} as const
export type EpisodesStatusOptions = typeof EpisodesStatusOptions[keyof typeof EpisodesStatusOptions]

export const EpisodesLivestreamOverrideOptions = {
	"inherit": "inherit",
	"live": "live",
	"skip": "skip",
} as const
export type EpisodesLivestreamOverrideOptions = typeof EpisodesLivestreamOverrideOptions[keyof typeof EpisodesLivestreamOverrideOptions]
export type EpisodesRecord = {
	cancelReason?: string
	createdBy?: RecordIdString
	endTime: IsoDateString
	genres?: RecordIdString[]
	hosts?: RecordIdString
	id: string
	image?: FileNameString
	isAutoGenerated?: boolean
	livestream_override?: EpisodesLivestreamOverrideOptions
	notes?: HTMLString
	programme?: RecordIdString
	show?: RecordIdString
	startTime: IsoDateString
	status: EpisodesStatusOptions
	title?: string
}